requirements.txt
fastapi[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
pytest
python-dotenv
